# sira-backend-master 2/swagger-api.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/leads": {
            "post": {
                "operationId": "LeadsController_create",
                "summary": "Create a new lead",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateLeadDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lead created successfully"
                    }
                },
                "tags": [
                    "leads"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "get": {
                "operationId": "LeadsController_findAll",
                "summary": "Get all leads with pagination and filters",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Leads retrieved successfully"
                    }
                },
                "tags": [
                    "leads"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/leads/bulk-assign": {
            "patch": {
                "operationId": "LeadsController_bulkAssign",
                "summary": "Bulk assign leads to a user",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BulkAssignLeadsDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Leads assigned successfully"
                    }
                },
                "tags": [
                    "leads"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/leads/bulk-rotation": {
            "patch": {
                "operationId": "LeadsController_bulkRotation",
                "summary": "Send leads to rotation pool (unassign, status rotation)",
                "description": "Rejected if any lead has at least one meeting (engagement).",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BulkSendToRotationDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Leads moved to rotation"
                    }
                },
                "tags": [
                    "leads"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/leads/batch-import/{batchId}": {
            "post": {
                "operationId": "LeadsController_importBatch",
                "summary": "Import leads from Excel/CSV for a specific batch",
                "parameters": [
                    {
                        "name": "batchId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "leads"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/leads/{id}": {
            "get": {
                "operationId": "LeadsController_findOne",
                "summary": "Get lead by ID",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lead found"
                    },
                    "404": {
                        "description": "Lead not found"
                    }
                },
                "tags": [
                    "leads"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "patch": {
                "operationId": "LeadsController_update",
                "summary": "Update lead",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateLeadDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lead updated successfully"
                    },
                    "404": {
                        "description": "Lead not found"
                    }
                },
                "tags": [
                    "leads"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "operationId": "LeadsController_remove",
                "summary": "Delete lead",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lead deleted successfully"
                    },
                    "404": {
                        "description": "Lead not found"
                    }
                },
                "tags": [
                    "leads"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/units/{id}": {
            "get": {
                "operationId": "UnitsController_findOne",
                "summary": "Get a unit by ID (sales: own or published only)",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Units"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "patch": {
                "operationId": "UnitsController_update",
                "summary": "Update unit inventory (not marketing)",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUnitDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Units"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "operationId": "UnitsController_remove",
                "summary": "Delete a unit (not marketing)",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Units"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/resale-units": {
            "post": {
                "operationId": "ResaleUnitsController_create",
                "summary": "Create a new resale unit",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateResaleUnitDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "Resale Units"
                ]
            },
            "get": {
                "operationId": "ResaleUnitsController_findAll",
                "summary": "Get all resale units",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Resale Units"
                ]
            }
        },
        "/resale-units/{id}": {
            "get": {
                "operationId": "ResaleUnitsController_findOne",
                "summary": "Get a resale unit by ID",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Resale Units"
                ]
            },
            "patch": {
                "operationId": "ResaleUnitsController_update",
                "summary": "Update a resale unit",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateResaleUnitDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Resale Units"
                ]
            },
            "delete": {
                "operationId": "ResaleUnitsController_remove",
                "summary": "Delete a resale unit",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Resale Units"
                ]
            }
        },
...
    },
    "info": {
        "title": "CRM API",
        "description": "Comprehensive CRM API for real estate management",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
...
            "CreateLeadDto": {
                "type": "object",
                "properties": {
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "leadSourceId": {
                        "type": "number"
                    },
                    "campaignId": {
                        "type": "number"
                    },
                    "dataBatchId": {
                        "type": "number"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "new",
                            "assigned",
                            "contacted",
                            "qualified",
                            "interested",
                            "not_interested",
                            "no_answer",
                            "converted",
                            "lost",
                            "rotation",
                            "another_meeting"
                        ],
                        "default": "new"
                    },
                    "priority": {
                        "type": "string",
                        "enum": [
                            "low",
                            "medium",
                            "high",
                            "urgent"
                        ],
                        "default": "medium"
                    },
                    "isStarred": {
                        "type": "boolean"
                    },
                    "assignedTo": {
                        "type": "number"
                    },
                    "assignmentMode": {
                        "type": "string",
                        "enum": [
                            "standard",
                            "customize"
                        ],
                        "default": "standard"
                    },
                    "teamId": {
                        "type": "number"
                    },
                    "propertyPreferences": {
                        "type": "object"
                    },
                    "notes": {
                        "type": "string"
                    },
                    "leadFront": {
                        "type": "boolean",
                        "default": true
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "primary",
                            "cold_call",
                            "campaign"
                        ],
                        "default": "primary"
                    }
                },
                "required": [
                    "phone"
                ]
            },
            "BulkAssignLeadsDto": {
                "type": "object",
                "properties": {
                    "leadIds": {
                        "example": [
                            1,
                            2,
                            3
                        ],
                        "description": "Array of lead IDs to assign",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "assignedTo": {
                        "type": "number",
                        "example": 4,
                        "description": "User ID to assign the leads to"
                    },
                    "assignmentMode": {
                        "type": "string",
                        "enum": [
                            "standard",
                            "customize"
                        ],
                        "default": "standard"
                    },
                    "assignmentDurationHours": {
                        "type": "number",
                        "description": "Optional: keep this assignment only for N hours; when it ends, assignmentExpireAction runs (rotation or backup_sales). Omit or 0 for a normal (no expiry) assignment.",
                        "minimum": 0,
                        "maximum": 8760
                    },
                    "assignmentExpireAction": {
                        "type": "string",
                        "enum": [
                            "rotation",
                            "backup_sales"
                        ]
                    },
                    "backupAssignUserId": {
                        "type": "number",
                        "description": "Required when assignmentExpireAction is backup_sales"
                    }
                },
                "required": [
                    "leadIds",
                    "assignedTo",
                    "assignmentMode"
                ]
            },
            "BulkSendToRotationDto": {
                "type": "object",
                "properties": {
                    "leadIds": {
                        "example": [
                            1,
                            2,
                            3
                        ],
                        "description": "Lead IDs to unassign and mark as rotation",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "required": [
                    "leadIds"
                ]
            },
            "UpdateLeadDto": {
                "type": "object",
                "properties": {
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "leadSourceId": {
                        "type": "number"
                    },
                    "campaignId": {
                        "type": "number"
                    },
                    "dataBatchId": {
                        "type": "number"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "new",
                            "assigned",
                            "contacted",
                            "qualified",
                            "interested",
                            "not_interested",
                            "no_answer",
                            "converted",
                            "lost",
                            "rotation",
                            "another_meeting"
                        ],
                        "default": "new"
                    },
                    "priority": {
                        "type": "string",
                        "enum": [
                            "low",
                            "medium",
                            "high",
                            "urgent"
                        ],
                        "default": "medium"
                    },
                    "isStarred": {
                        "type": "boolean"
                    },
                    "assignedTo": {
                        "type": "number"
                    },
                    "assignmentMode": {
                        "type": "string",
                        "enum": [
                            "standard",
                            "customize"
                        ],
                        "default": "standard"
                    },
                    "teamId": {
                        "type": "number"
                    },
                    "propertyPreferences": {
                        "type": "object"
                    },
                    "notes": {
                        "type": "string"
                    },
                    "leadFront": {
                        "type": "boolean",
                        "default": true
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "primary",
                            "cold_call",
                            "campaign"
                        ],
                        "default": "primary"
                    },
                    "assignmentDurationHours": {
                        "type": "number",
                        "description": "Temporary assignment window in hours (from assign time). 0 clears timer. Requires assignmentExpireAction when > 0.",
                        "minimum": 0,
                        "maximum": 8760
                    },
                    "assignmentExpireAction": {
                        "type": "string",
                        "enum": [
                            "rotation",
                            "backup_sales"
                        ]
                    },
                    "backupAssignUserId": {
                        "type": "number",
                        "description": "Required when assignmentExpireAction is backup_sales"
                    }
                }
            },
...
            "UpdateUnitDto": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "description": "Omit to auto-generate a unique code (recommended)",
                        "maxLength": 100
                    },
                    "description": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string",
                        "description": "Short unit / listing address line"
                    },
                    "projectName": {
                        "type": "string"
                    },
                    "location": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "monthlyInstallment": {
                        "type": "number",
                        "description": "Monthly installment"
                    },
                    "deliveryDate": {
                        "type": "string",
                        "description": "Delivery / handover date (YYYY-MM-DD)"
                    },
                    "bedrooms": {
                        "type": "number"
                    },
                    "bathrooms": {
                        "type": "number"
                    },
                    "floor": {
                        "type": "number",
                        "description": "Floor number (building level)"
                    },
                    "area": {
                        "type": "number",
                        "description": "Area in sqm or sqft"
                    },
                    "unitType": {
                        "type": "string"
                    },
                    "driveMediaLink": {
                        "type": "string",
                        "description": "Google Drive (or similar) link to media / documents — paste share link",
                        "maxLength": 1024
                    },
                    "amenities": {
                        "description": "Amenity labels (e.g. tags)",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "externalLinks": {
                        "description": "External URLs (listings, walkthroughs, etc.)",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "available",
                            "reserved",
                            "sold",
                            "unavailable"
                        ],
                        "default": "available"
                    }
                }
            },
...
        }
    }
}